# 🧠 Optimize `go` command parsing